requirements.txt
# requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
--find-links https://data.pyg.org/whl/torch-1.11.0+cu113.html
torch == 1.11.0+cu113
torchaudio == 0.11.0+cu113
torchvision == 0.12.0+cu113
torch-scatter
torch-sparse
# ---
addict
hydra-core
matplotlib
memory_profiler
neptune-client
numpy
omegaconf
pip
pytest
pytorch-lightning
scipy
sortednp
tqdm
